requirements.txt
Flask==2.0.1
Pillow==9.5.0
numpy==1.24.3
scipy==1.10.1
werkzeug==2.0.3
pytest==7.3.1
Flask-Testing==0.8.1
requests==2.31.0
